# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wesley's Ambacht Catering API
  description: |
    # Wesley's Ambacht Catering API

    A comprehensive REST API for managing catering bookings, quotes, availability, and add-on services for Wesley's Ambacht, a premium Dutch catering company.

    ## Features

    - **Real-time Availability Management** - Live booking slots with automatic reservation
    - **Dynamic Quote Generation** - Intelligent pricing based on service tiers and guest counts
    - **Add-on Services** - Flexible service extensions with per-person and flat-rate pricing
    - **Multi-tier Service Categories** - Essential, Premium, and Luxury tiers for different events
    - **Real-time Updates** - WebSocket subscriptions for live booking and availability updates
    - **Comprehensive Logging** - Full request/response logging with performance metrics

    ## Authentication

    This API uses Supabase Row Level Security (RLS) with JWT tokens. Include your authorization token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting

    - 100 requests per minute for general endpoints
    - 500 requests per minute for availability checks
    - 50 requests per minute for booking creation

    ## Error Handling

    All endpoints return structured error responses:

    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Guest count must be between 10 and 200",
        "details": {
...
        }
...

...
servers:
...
paths:
...
  /availability/check:
    post:
      summary: Check specific slot availability
      description: |
        Check if a specific date and time slot is available for booking.

        **Real-time**: This endpoint provides real-time availability status.
      operationId: checkAvailability
      tags:
        - Availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - time
              properties:
                date:
                  type: string
                  format: date
                  description: Date to check (YYYY-MM-DD)
                  example: "2024-03-15"
                time:
                  type: string
                  pattern: "^[0-2][0-9]:[0-5][0-9]$"
                  description: Time slot to check (HH:MM)
                  example: "18:00"
      responses:
        "200":
          description: Availability status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                  remaining_slots:
                    type: integer
                  time_slot:
                    type: string
              example:
                available: true
                remaining_slots: 2
                time_slot: "18:00"
...
  /bookings:
    get:
      summary: List bookings
      description: |
        Retrieve a list of bookings with filtering and pagination options.

        **Security**: Only returns bookings accessible to the authenticated user.
      operationId: listBookings
      tags:
        - Bookings
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          description: Filter by booking status
          schema:
            $ref: "#/components/schemas/BookingStatus"
        - name: service_category
          in: query
          description: Filter by service category
          schema:
            $ref: "#/components/schemas/ServiceCategory"
        - name: date_from
          in: query
          description: Filter bookings from this date
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: Filter bookings to this date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

    post:
      summary: Create a new booking
      description: |
        Create a new catering booking. Automatically checks availability and reserves the time slot.

        **Validation**: All fields are validated according to business rules.
        **Side Effects**: Reserves the selected time slot, sends confirmation email.
      operationId: createBooking
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
                  message:
                    type: string
                    example: "Booking created successfully"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Time slot no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "SLOT_UNAVAILABLE"
                  message: "The selected time slot is no longer available"
  /bookings/{id}:
    get:
      summary: Get booking details
      description: |
        Retrieve detailed information about a specific booking.
      operationId: getBooking
      tags:
        - Bookings
      parameters:
        - name: id
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update booking status
      description: |
        Update the status of a booking. Some status changes trigger side effects.

        **Side Effects**:
        - `confirmed`: Sets confirmed_at timestamp, sends confirmation email
        - `cancelled`: Releases time slot, sets cancelled_at timestamp
      operationId: updateBookingStatus
      tags:
        - Bookings
      parameters:
        - name: id
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: "#/components/schemas/BookingStatus"
                notes:
                  type: string
                  description: Optional notes about the status change
                  maxLength: 500
      responses:
        "200":
          description: Booking status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Booking"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "NOT_FOUND"
              message: "The requested resource was not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Validation failed"
              details:
                field: "guest_count"
                value: 5
                message: "Guest count must be between 10 and 200"
...
  schemas:
    ServiceCategory:
      type: string
      enum: [corporate, private, wedding, celebration]
      description: |
        Type of catering service:
        - `corporate`: Business events and corporate catering
        - `private`: Private parties and intimate gatherings  
        - `wedding`: Wedding ceremonies and receptions
        - `celebration`: Special celebrations and milestone events
    ServiceTier:
      type: string
      enum: [essential, premium, luxury]
      description: |
        Service level tiers:
        - `essential`: Basic service package with quality essentials
        - `premium`: Enhanced service with premium options
        - `luxury`: Top-tier service with luxury amenities
    BookingStatus:
      type: string
      enum: [pending, confirmed, cancelled, completed]
      description: |
        Booking status lifecycle:
        - `pending`: Initial booking, awaiting confirmation
        - `confirmed`: Booking confirmed and scheduled
        - `cancelled`: Booking cancelled (releases time slot)
        - `completed`: Event completed successfully
...
    Booking:
      type: object
      required:
        - id
        - customer_name
        - customer_email
        - event_date
        - event_time
        - guest_count
        - service_category
        - service_tier
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        customer_name:
          type: string
          maxLength: 100
          description: Full name of the customer
        customer_email:
          type: string
          format: email
          maxLength: 254
          description: Customer's email address
        customer_phone:
          type: string
          maxLength: 20
          nullable: true
          description: Customer's phone number
        company_name:
          type: string
          maxLength: 100
          nullable: true
          description: Company name for corporate events
        event_date:
          type: string
          format: date
          description: Date of the event
        event_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
          description: Start time of the event
        guest_count:
          type: integer
          minimum: 10
          maximum: 500
          description: Number of guests expected
        service_category:
          $ref: "#/components/schemas/ServiceCategory"
        service_tier:
          $ref: "#/components/schemas/ServiceTier"
        status:
          $ref: "#/components/schemas/BookingStatus"
        special_requests:
          type: string
          maxLength: 1000
          nullable: true
          description: Any special requests or notes
        dietary_restrictions:
          type: string
          maxLength: 500
          nullable: true
          description: Dietary restrictions or allergies
        estimated_total:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Estimated total cost in EUR
        final_total:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Final confirmed total cost in EUR
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
    CreateBookingRequest:
      type: object
      required:
        - customer_name
        - customer_email
        - event_date
        - event_time
        - guest_count
        - service_category
      properties:
        customer_name:
          type: string
          maxLength: 100
          minLength: 2
          example: "Jan van der Berg"
        customer_email:
          type: string
          format: email
          maxLength: 254
          example: "jan@example.com"
        customer_phone:
          type: string
          maxLength: 20
          pattern: '^[+]?[0-9\s\-\(\)]{10,}$'
          example: "+31 6 12345678"
        company_name:
          type: string
          maxLength: 100
          example: "ABC Consulting B.V."
        event_date:
          type: string
          format: date
          description: Must be in the future
          example: "2024-03-15"
        event_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
          example: "18:00"
        guest_count:
          type: integer
          minimum: 10
          maximum: 500
          example: 25
        service_category:
          $ref: "#/components/schemas/ServiceCategory"
        service_tier:
          $ref: "#/components/schemas/ServiceTier"
          default: "premium"
        special_requests:
          type: string
          maxLength: 1000
          example: "Vegetarian options preferred, outdoor seating if weather permits"
        dietary_restrictions:
          type: string
          maxLength: 500
          example: "2 guests with nut allergies, 1 vegan"
        selected_add_ons:
          type: array
          items:
            type: string
            format: uuid
          description: Array of add-on service IDs
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
        total_pages:
          type: integer
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there are more pages
        has_prev:
          type: boolean
          description: Whether there are previous pages
security:
  - BearerAuth: []
...